publishing

+++ src/TejasCaptchaSessionCleanup.php
@@ -27,11 +27,25 @@ class TejasCaptchaSessionCleanup {
   protected $path;
 
   /**
+   * The path where final audio files are stored.
+   *
+   * @var string
+   */
+  protected $filenames;
+
+  /**
    * The number of minutes the session audio should be valid.
    *
    * @var int
    */
   protected $minutes;
+
+  /**
+   * The number of files in the audio directory.
+   *
+   * @var int
+   */
+  protected $filecount;
 
     /**
     * Create a new TejasCaptcha garbage collector
@@ -39,6 +53,9 @@ class TejasCaptchaSessionCleanup {
     public function __construct()
     {
       $this->filesystem = new Filesystem();
+      $this->path = 'var/www/dev.173.255.195.42/resources/audio';
+      $this->filenames = null;
+      $this->minutes = 1;
     }
 
     /**
@@ -48,19 +65,54 @@ class TejasCaptchaSessionCleanup {
     */
     public function gc(array $options = []) {
         $this->path = $options['path'] ?? $this->path;
-        $this->minutes = $options['minutes'] ?? $this->minutes;
-
+        $this->filenames = $options['filenames'] ?? $this->filenames;
         try{
             $files = Finder::create()
                         ->in($this->path)
                         ->files()
                         ->ignoreDotFiles(true)
                         ->date('<= now - '.$this->minutes.' minutes');
-            foreach ($files as $file) {
-                $this->filesystem->delete($file->getRealPath());
-            }
+
+              foreach ($files as $file) {
+                  $this->filesystem->delete($file->getRealPath());
+              }
+
         }catch(exception $e){
             LOG::debug('Caught exception: ' . $e->getMessage() . "\n");
         }
+
+        try{
+          $files = Finder::create()
+                      ->in($this->path)
+                      ->files()
+                      ->ignoreDotFiles(true);
+
+            if(iterator_count($files) > 50) {
+                foreach ($files as $file) {
+                    $this->filesystem->delete($file->getRealPath());
+                }
+            }
+
+        } catch(exception $e){
+            LOG::debug('Caught exception: ' . $e->getMessage() . "\n");
+        }
+
+
+        if($this->filenames !== null) {
+            try{
+                $files = Finder::create()
+                            ->in($this->path)
+                            ->name($this->filenames);
+
+                foreach ($files as $file) {
+                    $this->filesystem->delete($file->getRealPath());
+                }
+
+              }catch(exception $e){
+                  LOG::debug('Caught exception: ' . $e->getMessage() . "\n");
+              }
+          }
+
     }
+
 }
